backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic[email]==2.5.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0